requirements.txt
# HTTP requests and networking
requests>=2.31.0
urllib3>=2.0.7

# Data parsing and handling
PyYAML>=6.0.1
beautifulsoup4>=4.12.2
networkx>=3.1

# Terminal output formatting
colorama>=0.4.6

# API and documentation handling
gql>=3.5.0
graphql-core>=3.2.3

# Type hints and utilities
typing-extensions>=4.7.1

# Data analysis
pandas>=2.0.0

# Development dependencies
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
